editor/libeditor/tests/test_bug622371.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=622371
-->
<head>
  <title>Test for Bug 622371</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="/tests/SimpleTest/EventUtils.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=622371">Mozilla Bug 622371</a>
<p id="display"></p>
<div id="content">
  <iframe srcdoc="<body contenteditable>abc</body>"></iframe>
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 622371 **/
SimpleTest.waitForExplicitFinish();
addLoadEvent(function() {
  var i = document.querySelector("iframe");
  var sel = i.contentWindow.getSelection();
  var doc = i.contentDocument;
  var body = doc.body;
  i.focus();
  sel.collapse(body, 1);
  doc.designMode = "on";
  doc.designMode = "off";
  is(sel.getRangeAt(0).startOffset, 1, "The start offset of the selection shouldn't change");
  is(sel.getRangeAt(0).endOffset, 1, "The end offset of the selection shouldn't change");
  SimpleTest.finish();
});

</script>
</pre>
</body>
</html>
